% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_time_debug.R
\name{debug_time_plot}
\alias{debug_time_plot}
\title{Interactive time metric comparison plot}
\usage{
debug_time_plot(
  nif,
  analyte = NULL,
  dose = NULL,
  color = NULL,
  facet = "DOSE",
  min_time = NULL,
  max_time = NULL,
  size = 2,
  alpha = 0.5,
  log = FALSE,
  title = "Time metric relation",
  legend = TRUE,
  scales = "fixed",
  ...
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte(s) to be plotted, as character.}

\item{dose}{The dose(s) to be filtered for.}

\item{color}{The column(s) to be used for coloring.}

\item{facet}{The column(s) to be used for faceting.}

\item{min_time}{The minimal time in units of the selected time field, as
numeric.}

\item{max_time}{The minimal time in units of the selected time field, as
numeric.}

\item{size}{The \code{size} parameter to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} as numeric.}

\item{alpha}{The \code{alpha} parameter to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, as numeric.}

\item{log}{Logarithmic y axis, as logical.}

\item{title}{The plot title, as character. If none is provided, a generic
title based on the analyte and the faceting will be chosen. Override this
by setting title = "", if needed.}

\item{legend}{Show the plot legend, as logical.}

\item{scales}{The \code{scales} parameter to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}, can be
"fixed" (default), "free", "free_x" or "free_y".}

\item{...}{Further parameters.}
}
\value{
A Shiny app object.
}
\description{
This function creates an interactive Shiny plot to compare different time
metrics (TIME, TAFD, TAD, NTIME) against each other. This is useful for
debugging and verifying the consistency of time-related fields in a NIF data
set. The x- and y-axis time metrics can be selected interactively via
dropdown menus.
}
\examples{
if (interactive()) {
  debug_time_plot(examplinib_poc_nif)
  debug_time_plot(examplinib_poc_nif, color = "ANALYTE")
  debug_time_plot(examplinib_poc_nif, analyte = "RS2023", facet = NULL)
}
}
