% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_debug_plot.R
\name{nif_debug}
\alias{nif_debug}
\title{Interactive debug plot for NIF objects}
\usage{
nif_debug(
  nif,
  sdtm,
  analyte = NULL,
  dose = NULL,
  time = "TAFD",
  color = NULL,
  facet = "DOSE",
  min_time = NULL,
  max_time = NULL,
  cfb = FALSE,
  dose_norm = FALSE,
  log = FALSE,
  lines = TRUE,
  size = 2,
  alpha = 1,
  scales = "fixed"
)
}
\arguments{
\item{nif}{A nif object containing `SRC_DOMAIN` and `SRC_SEQ` columns.}

\item{sdtm}{A sdtm object used to look up source records.}

\item{analyte}{The analyte(s) to be plotted, as character.}

\item{dose}{The dose(s) to be filtered for.}

\item{time}{The time field for the x-axis. One of 'TIME', 'NTIME', 'TAFD'
or 'TAD'.}

\item{color}{The column(s) to be used for coloring.}

\item{facet}{The column(s) to be used for faceting.}

\item{min_time}{The minimal time, as numeric.}

\item{max_time}{The maximal time, as numeric.}

\item{cfb}{Plot change from baseline, as logical.}

\item{dose_norm}{Dose-normalized values, as logical.}

\item{log}{Logarithmic y axis, as logical.}

\item{lines}{Plot lines, as logical.}

\item{size}{The `size` parameter to [ggplot2::geom_point()], as numeric.}

\item{alpha}{The `alpha` parameter to [ggplot2::geom_point()], as numeric.}

\item{scales}{The `scales` parameter to [ggplot2::facet_wrap()].}
}
\value{
A Shiny app object.
}
\description{
Launches a Shiny app that renders an interactive version of [nif::plot.nif()].
Data points are clickable; clicking a point displays the corresponding
source SDTM record (identified via `SRC_DOMAIN` and `SRC_SEQ`) in a table
below the plot.
}
\details{
The NIF object must contain `SRC_DOMAIN` and `SRC_SEQ` columns. These are
present when the NIF is built with `debug = TRUE` or when the global
`nif::nif_option("debug")` is set.
}
\seealso{
[nif::plot.nif()], [nif::nif_viewer()]
}
